small changes added

+++ Gui/GuiMainView.py
@@ -1,3 +1,4 @@
+import shutil
 import string
 from tkinter import *
 from tkinter import filedialog
@@ -6,41 +7,45 @@ from threading import Thread
 
 import os
 
+import math
+
 from BasicMethods import get2DArrayFromFile
 from Gui.TkinterTable import TableView
 
 
 class EngineBuilder(Frame):
 
-    def __init__(self, master, mainManager, config, numOfTotalFiles):
+    def __init__(self, master, mainManager, config):
         self.config = config
         self.mainManager = mainManager
         Frame.__init__(self, master)
         self.grid()
         # self.filesDone = 0
-        self.numOfTotalFiles = numOfTotalFiles
         self.numOfFilesPerIteration = config.get__filesPerIteration()
 
+        self.XstartPixel = 60
+        self.YstartPixel = 10
 
-        label_0 = Label(self.master, text="Search Engine", width=20, font=("bold", 20))
-        label_0.place(x=90, y=60)
+
+        label_0 = Label(self.master, text="Search Engine", width=20, font=("bold", 30))
+        label_0.place( x = self.XstartPixel + 20, y = self.YstartPixel + 40)
 
 
 
         label_corpusPath = Label(self.master, text="Corpus path:", width=10, font=("bold", 10))
-        label_corpusPath.place(x=50, y=130)
+        label_corpusPath.place( x = self.XstartPixel + 50, y = self.YstartPixel + 130)
         self.entry_corpusPath_text = StringVar()
         self.entry_corpusPath_text.set(config.get__corpusPath())
         self.entry_corpusPath = Entry(self.master,textvariable=self.entry_corpusPath_text,width=30)
-        self.entry_corpusPath.place(x=180, y=130)
+        self.entry_corpusPath.place( x = self.XstartPixel + 180, y = self.YstartPixel + 130)
 
 
         label_postingPath = Label(self.master, text="Posting path:", width=10, font=("bold", 10))
-        label_postingPath.place(x=50, y=160)
+        label_postingPath.place( x = self.XstartPixel + 50,  y = self.YstartPixel + 160)
         self.entry_postingPath_text = StringVar()
         self.entry_postingPath_text.set(config.get__savedFileMainFolder())
         self.entry_postingPath = Entry(self.master,textvariable=self.entry_postingPath_text,width=30)
-        self.entry_postingPath.place(x=180, y=160)
+        self.entry_postingPath.place( x = self.XstartPixel + 180, y = self.YstartPixel + 160)
 
 
         def corpusPath():
@@ -58,20 +63,21 @@ class EngineBuilder(Frame):
 
 
         self.corpusPathButton = Button(self.master, text='Find', width=5, fg='black',command= corpusPath)
-        self.corpusPathButton.place(x=380, y=125)
+        self.corpusPathButton.place( x = self.XstartPixel + 380, y = self.YstartPixel + 125)
         self.postingPathButton = Button(self.master, text='Find', width=5, fg='black',command= postingPath)
-        self.postingPathButton.place(x=380, y=155)
+        self.postingPathButton.place( x = self.XstartPixel + 380, y = self.YstartPixel + 155)
 
 
 
-        self.data = None
+        self.dataStem = None
+        self.dataNoStem = None
         self.headline = None
 
 
 
 
-        label_4 = Label(self.master, text="Language", width=10, font=("bold", 10))
-        label_4.place(x=50, y=190)
+        label_4 = Label(self.master, text="Language:", width=10, font=("bold", 10))
+        label_4.place( x = self.XstartPixel + 50, y = self.YstartPixel + 190)
 
 
         list1 = ['English', 'Spanish', 'Hebrew']
@@ -79,91 +85,167 @@ class EngineBuilder(Frame):
         self.droplist = OptionMenu(self.master, c, *list1)
         self.droplist.config(width=15)
         c.set('Select')
-        self.droplist.place(x=180, y=190)
+        self.droplist.place( x = self.XstartPixel + 180, y = self.YstartPixel + 190)
 
 
 
 
         self.deleteButton = Button(self.master, text='Delete', width=10, bg='red', fg='white',command= self.deleteEngine)
-        self.deleteButton.place(x=100, y=250)
+        self.deleteButton.place( x = self.XstartPixel + 100, y = self.YstartPixel + 250)
         self.buildButton = Button(self.master, text='Build', width=10, bg='green', fg='white',command= self.buildEngine)
-        self.buildButton.place(x=200, y=250)
+        self.buildButton.place( x = self.XstartPixel + 200, y = self.YstartPixel + 250)
 
 
         label_stemming = Label(self.master, text="Stemming", width=10, font=("bold", 10))
-        label_stemming.place(x=320, y=250)
+        label_stemming.place( x = self.XstartPixel + 320, y = self.YstartPixel + 250)
         self.checked = BooleanVar()
         self.stemmingCheckBox = Checkbutton(self.master, variable = self.checked)
-        self.stemmingCheckBox.place(x=300, y=250)
+        self.stemmingCheckBox.place( x = self.XstartPixel + 300, y = self.YstartPixel + 250)
 
 
-        self.label_progress = Label(self.master, text="Progress:     [||||||||||||||||||||||||||||||||||||||||          ] 80% ", width=50, font=("bold", 10))
-        self.label_progress.place(x=50, y=320)
+        # Create progress bar
+
+        self.label_postingProgress = Label(self.master, text=" ", width=50, font=("bold", 10))
+        self.label_postingProgress.place(x =self.XstartPixel + 40, y =self.YstartPixel + 300)
+
+        self.label_mergeProgress = Label(self.master, text=" ", width=50, font=("bold", 10))
+        self.label_mergeProgress.place(x =self.XstartPixel + 40, y =self.YstartPixel + 320)
+
+
+
+        # Set Progress bar
+        linesAsString = ''
+        for i in range(0, 51):
+            if i % 4 == 0:
+                continue
+            linesAsString += ' '
+        self.posting_progressLabel = "Posting Progress:     ["
+        self.merge_progressLabel = "Merge Progress:       ["
+        self.label_postingProgress['text'] = self.posting_progressLabel + linesAsString + '] 00%'
+        self.label_mergeProgress['text'] = self.merge_progressLabel + linesAsString + '] 00%'
 
 
 
         label_postingPath = Label(self.master, text="Dictionary:", width=20, font=("bold", 10))
-        label_postingPath.place(x=30, y=380)
+        label_postingPath.place( x = self.XstartPixel + 30, y = self.YstartPixel + 380)
 
 
 
         self.uploadDicButton = Button(self.master, text='Upload', width=10, bg='blue', fg='white',command= self.loadDictionary)
-        self.uploadDicButton.place(x=170, y=380)
+        self.uploadDicButton.place( x = self.XstartPixel + 170, y = self.YstartPixel + 380)
         self.showDicButton = Button(self.master, text='Show', width=10, bg='blue', fg='white',command= self.displayDicionary)
-        self.showDicButton.place(x=270, y=380)
+        self.showDicButton.place(x = self.XstartPixel + 270, y = self.YstartPixel + 380)
 
         self.label_buildDetails = Label(self.master, text="",width=50 ,font=("bold",10))
-        self.label_buildDetails.place(x=50,y=420)
+        self.label_buildDetails.place( x = self.XstartPixel + 50, y = self.YstartPixel + 420)
 
 
-        Label(self.master, text="Summary:", width=10, font=("bold", 10)).place(x=20, y=420)
+        Label(self.master, text="Summary:", width=10, font=("bold", 10)).place( x = self.XstartPixel + 20, y = self.YstartPixel + 420)
 
 
         from tkinter import scrolledtext
-        self.txtbox = scrolledtext.ScrolledText(width= 45,height=7)
-        self.txtbox.place(x= 60, y = 440)
+        self.txtbox = scrolledtext.ScrolledText( width = 45, height = 10)
+        self.txtbox.place( x = self.XstartPixel + 60, y = self.YstartPixel + 450)
 
-        self.statusLabel = Label(self.master, text="Status: Ready to Build\Shut down", width=40, font=("bold", 10))
-        self.statusLabel.place(x=60, y=570)
-
-
+        self.statusLabel = Label(self.master, text="Status: Ready to Build\Shut down", width = 40, font = ("bold", 10))
+        self.statusLabel.place( x = self.XstartPixel + 60, y = self.YstartPixel + 650)
 
 
 
 
 
 
-    def updatePostingProgress(self):
+
+
+    def updateProgress(self, posting_merge):
         flag = True
-        label = "Progress:     ["
         import os
         import time
-        counter = 0
+
+
+        sleepTime = 2
 
         while flag:
-            time.sleep(10)
-            path = self.config.get__savedFilePath() + '/Progress/Posting'
+
+            time.sleep(sleepTime)
+            counter = 0
+
+            path = self.config.get__savedFilePath() + '/Progress/%s' % (posting_merge)
             if not os.path.exists(path):
                 break
             listOfFiles = os.listdir(path)
             if len(listOfFiles) == 0:
                 continue
-            for file in listOfFiles:
-                splitedFile = file.split('_')
-                counter += int(splitedFile[-1])
 
-            if counter == self.numOfTotalFiles:
+            dicByManagerID = {}
+            lastCount = ''
+            for file in sorted(listOfFiles):
+                splitedFile = file.split('_')
+                managerID = splitedFile[0]
+                managerFileCount = int(splitedFile[1])
+
+                if managerID == '-1':
+                    # sleepTime = 10
+                    shutil.rmtree(path)
+                    os.mkdir(path)
+                    dicByManagerID = {}
+                    dicByManagerID['-1'] = managerFileCount
+                    break
+                if dicByManagerID.get(managerID) is None:
+                    dicByManagerID[managerID] = managerFileCount
+                else:
+                    if managerFileCount > dicByManagerID[managerID]:
+                        dicByManagerID[managerID] = managerFileCount
+
+                lastCount = str(managerFileCount)
+
+            for value in dicByManagerID.values():
+                counter += value
+
+
+
+            if posting_merge == 'Posting':
+                percent = (counter / self.config.get_numOfFiles()) * 50
+                percent = int(percent)
+
+            elif posting_merge == 'Merge':
+                # print(dicByManagerID)
+                numOfFiles = counter
+                filesPerIteration = 10
+                filesPerIterationMerge = 10
+                numOfManagers = self.config.managersNumber
+
+                # F + [F/N] + [(F/N)/I] + [(F/N)(I/M)] + 27*10
+                # percent = numOfFiles + math.ceil(numOfFiles/numOfManagers) + math.ceil((numOfFiles/numOfManagers)/filesPerIteration) + math.ceil((numOfFiles/numOfManagers)/(filesPerIteration/filesPerIterationMerge)) + 27*10
+
+                percent = (counter / (self.config.get_numOfFiles() + 27*50)) * 50
+                percent = int(percent)
+
+
+
+            if percent >= 49:
+                linesAsString = ''
+                for i in range(0, 51):
+                    linesAsString += '|'
+                self.label_postingProgress['text'] = self.posting_progressLabel + linesAsString + '] 100%'
                 return
-            percent = (counter/self.numOfTotalFiles)*50
-            percent = int(percent)
+
             linesAsString = ''
-            for i in range(0,percent):
+            for i in range(0,percent + 1):
                 linesAsString += '|'
-            for i in range(percent,50):
+            for i in range(percent + 1,51):
+                if i % 4 == 0:
+                    continue
                 linesAsString += ' '
 
-            self.label_progress['text'] = label + linesAsString + '] ' + str(percent*2) + '% '
+            percentString = str(percent*2)
+            if len(percentString) == 1:
+                percentString = '0' + percentString
 
+            if posting_merge == 'Posting':
+                self.label_postingProgress['text'] = self.posting_progressLabel + linesAsString + '] ' + percentString + '% '
+            elif posting_merge == 'Merge':
+                self.label_mergeProgress['text'] = self.merge_progressLabel + linesAsString + '] ' + percentString + '% '
 
 
 
@@ -174,6 +256,7 @@ class EngineBuilder(Frame):
         lettersList = list(string.ascii_lowercase)
         lettersList.append('#')
 
+
         totalList = []
         for letter in lettersList:
             path = savedFolderPath + '/' + letter + '/' + 'mergedFile_dic'
@@ -182,10 +265,20 @@ class EngineBuilder(Frame):
                 self.statusLabel['text'] = 'Need to build (Check if stem is clicked)'
                 print('Location not found', path)
                 return
-            totalList = totalList + get2DArrayFromFile(path)
+            arrayFromFile = get2DArrayFromFile(path=path)
+            totalList = totalList + arrayFromFile
+
 
         self.headline = ['Term', 'df', 'sumTF', '# Posting']
-        self.data = totalList
+        if self.config.toStem:
+            self.dataStem = totalList
+        else:
+            self.dataNoStem = totalList
+
+
+
+
+
 
 
     def loadDictionary(self):
@@ -205,13 +298,19 @@ class EngineBuilder(Frame):
         self.config.setToStem(check)
 
         if not os.path.exists(self.config.savedFilePath + '/a'):
-            self.statusLabel['text'] = 'Status: Please build before trying to load'
+            if check:
+                self.statusLabel['text'] = 'Status (stem is checked): build before load'
+            else:
+                self.statusLabel['text'] = 'Status (stem not checked): build before load'
+
             return
 
 
 
 
         self.disableButtons()
+        self.statusLabel['text'] = 'Status: Loading terms'
+
         print('Load dictionary')
 
         t = Thread(target=self.load, args=())
@@ -224,16 +323,32 @@ class EngineBuilder(Frame):
 
     def displayDicionary(self):
 
+        check = self.checked.get()
+        self.config.setToStem(check)
+        data = None
+
+        if self.config.toStem:
+            if self.dataStem is None or self.headline is None:
+                self.statusLabel['text'] = 'Status (stem is checked): upload before Show'
+                return
+            else:
+                data = self.dataStem
+
+        else:
+            if self.dataNoStem is None or self.headline is None:
+                self.statusLabel['text'] = 'Status (stem not checked): upload before Show'
+                return
+            else:
+                data = self.dataNoStem
+
+
+
         self.disableButtons()
         print('Display dictionary')
 
-        if self.data is None or self.headline is None:
-            self.enableButtons()
-            return
+        self.statusLabel['text'] = 'Status: preparing a nice table to view terms'
 
-
-        self.displayClass = TableView(self.data, self.headline)
-
+        self.displayClass = TableView(data, self.headline)
 
 
 
@@ -245,12 +360,21 @@ class EngineBuilder(Frame):
 
     def deleteEngine(self):
         import shutil
+        saveMainFolderPath = str(self.entry_postingPath.get())
+        if not os.path.exists(saveMainFolderPath + '/SavedFiles'):
+            self.statusLabel['text'] = 'Status: %s invalid path to delete' % (saveMainFolderPath,)
+            return
+
+
         shutil.rmtree(self.config.get__savedFileMainFolder())
         print("Folder was deleted successfully..")
         self.buildButton.configure(state=NORMAL)
         self.deleteButton.configure(state=DISABLED)
         self.showDicButton.configure(state = DISABLED)
         self.uploadDicButton.configure(state = DISABLED)
+
+        self.statusLabel['text'] = 'Status: Folder %s was deleted' % (self.config.get__savedFileMainFolder())
+
 
 
 
@@ -289,7 +413,9 @@ class EngineBuilder(Frame):
 
         print("\n***    ManagerRun    ***\n")
 
-        self.statusLabel['text'] = 'Building.. might take a while'
+        from datetime import datetime
+        time = datetime.now().strftime('%H:%M:%S')
+        self.statusLabel['text'] = '%s - Started building.. might take a while' % (time)
 
         from threading import Thread
         from concurrent.futures import ThreadPoolExecutor
@@ -302,8 +428,12 @@ class EngineBuilder(Frame):
         threadWaitUntilBuildDone.start()
 
 
-        # threadProgress = Thread(target=updateFileCounter)
-        # threadProgress.start()
+        threadPostingProgress = Thread(target=self.updateProgress , args=('Posting',))
+        threadPostingProgress.start()
+
+
+        threadMergeProgress = Thread(target=self.updateProgress, args=('Merge',))
+        threadMergeProgress.start()
 
 
     @ staticmethod
@@ -315,20 +445,23 @@ class EngineBuilder(Frame):
 
 
     def buildListener(self,future):
+        from Parsing.ConvertMethods import converSecondsToTime
+
         timeItTook, maxParsingTime, totalMerging, totalNumberOfTerms, totalNumberOfDocuments ,mergedLanguagesSet = future.result()
         print("Number of Terms: " , str(totalNumberOfTerms))
         print("Number of Docs: " , str(totalNumberOfDocuments))
-        print("Parsing Time: " , str(maxParsingTime))
-        print("Merging Time: " , str(totalMerging))
-        print("Everything took: " , str(timeItTook) , " seconds")
+        print("Parsing Time: " , str(converSecondsToTime(maxParsingTime)))
+        print("Merging Time: " , str(converSecondsToTime(totalMerging)))
+        print("Everything took: " , str(converSecondsToTime(timeItTook)))
         self.enableButtons()
+
 
         self.droplist.destroy()
         c = StringVar()
         self.droplist = OptionMenu(self.master, c, *mergedLanguagesSet)
         self.droplist.config(width=15)
         c.set('Select')
-        self.droplist.place(x=180, y=190)
+        self.droplist.place( x = self.XstartPixel + 180, y = self.YstartPixel + 190)
         self.setBuildDetails(timeItTook, maxParsingTime, totalMerging, totalNumberOfTerms, totalNumberOfDocuments)
 
 
@@ -336,6 +469,7 @@ class EngineBuilder(Frame):
 
 
     def setBuildDetails(self, timeItTook, maxParsingTime, totalMerging, totalNumberOfTerms, totalNumberOfDocuments):
+        from Parsing.ConvertMethods import converSecondsToTime
         stem = ''
         if self.config.toStem:
             stem = '\n** Run with stemming - Details **\n'
@@ -344,9 +478,16 @@ class EngineBuilder(Frame):
         detailString = stem
         detailString += "\tNumber of Terms:  " + str(totalNumberOfTerms) + "\n"
         detailString += "\tNumber of Docs:   " + str(totalNumberOfDocuments) + "\n"
-        detailString += "\tParsing Time:     " + str(maxParsingTime) + "\n"
-        detailString += "\tMerging Time:     " + str(totalMerging) + "\n"
-        detailString += "\tEverything took:  " + str(timeItTook) + " seconds\n"
+        detailString += "\tParsing Time:     " + str(converSecondsToTime(maxParsingTime)) + "\n"
+        detailString += "\tMerging Time:     " + str(converSecondsToTime(totalMerging)) + "\n"
+        detailString += "\tEverything took:  " + str(converSecondsToTime(timeItTook)) + "\n"
+
+        linesAsString = ''
+        for i in range(0, 51):
+            linesAsString += '|'
+        self.label_postingProgress['text'] = self.posting_progressLabel + linesAsString + '] 100%'
+        self.label_mergeProgress['text'] = self.merge_progressLabel + linesAsString + '] 100%'
+
         self.txtbox.insert('end',detailString)
         # self.label_buildDetails['text'] = detailString
